SHARE-342 Remix + Code stats/view lazy loading
  ### Project page

- removed embedded remix graph, code stats, code view and remix count to highly reduce loading times

- program.html.twig major refactoring to reduce code duplication

- new design (mobile)

![image](https://user-images.githubusercontent.com/40868718/87135108-e53d3500-c299-11ea-85d4-081ecdb322e4.png)

- new design (large screens)

![image](https://user-images.githubusercontent.com/40868718/87135186-000fa980-c29a-11ea-8370-081600013674.png)

- is very computation intensive in its current implementation. Since most users are not interested in this graph, the feature was moved to it's own page.

- top bar is on top of the "full screen" remix graph, providing means to always return to the project page.

- The graph lazy loads the number of remixes in the title to improve performance

![image](https://user-images.githubusercontent.com/40868718/87134338-d7d37b00-c298-11ea-9e12-c132779f2191.png)

- are also very expensive and not used by many users, therefore moved to their own pages.

- both new pages provide top bar actions to return back to the project page and have an adapted top bar title

![image](https://user-images.githubusercontent.com/40868718/87135359-3cdba080-c29a-11ea-9c58-75c44e10d30f.png)

![image](https://user-images.githubusercontent.com/40868718/87135413-52e96100-c29a-11ea-8abb-2be74c68bdea.png)

- adapt/add tests
- make translation not uppercase
- change/add translations
