+++ assets/css/base/_base.scss
@@ -55,6 +55,7 @@ $fa-font-path: "../../webfonts";
 @import "catro_bootstrap";
 @import "brain";
 @import "icons";
+@import "buttons";
 
 // new design
 @import "typography";
