+++ assets/css/base/_buttons.scss
@@ -1,0 +1,6 @@
+.btn,
+.button {
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  overflow: hidden;
+}
