--- assets/css/base/_buttons_.scss
@@ -1,4 +1,0 @@
-button,
-.button {
-  color: #fff;
-}
