+++ assets/css/base/_icons.scss
@@ -42,11 +42,6 @@
   font-size: 1.2em;
 }
 
-.remix-close-icon
-{
-  font-size: 3em;
-}
-
 .like-round-icon-container
 {
   background-color: $primary;
