+++ assets/css/custom/loadinganimation.scss
@@ -3,7 +3,7 @@
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
-  z-index: 99999;
+  z-index: 600;
 }
 
 .loader {
