+++ assets/css/custom/remixgraph.scss
@@ -2,35 +2,12 @@
 
 #remix-graph-modal {
   margin: 0;
+  z-index: 300;
 }
 
 #remix-graph-layer {
   width: 100%;
   height: 100vh;
-}
-
-.close-remix-graph-modal {
-  position: absolute;
-  top: 0;
-  right: 0;
-  z-index: 1;
-  width: 70px;
-  height: 70px;
-  margin: 22px 18px 0;
-  text-align: left;
-}
-
-.close-button {
-  width: 70px;
-  height: 70px;
-  border: none;
-  outline: none;
-  transition: all 0.2s;
-  @include user-select(none);
-
-  &:hover {
-    transform: rotate(90deg);
-  }
 }
 
 .overlay {
@@ -39,7 +16,7 @@
   right: 0;
   bottom: 0;
   left: 0;
-  z-index: 20000;
+  z-index: 500;
   width: 100%;
   height: 100%;
   background-color: rgba(0, 0, 0, 0.01);
