+++ assets/js/analytics/analytics.js
@@ -1370,7 +1370,7 @@ if ((window.doNotTrack || navigator.doNotTrack || navigator.msDoNotTrack || 'msT
             /**
              * Tracks the display of the remix graph by link
              */
-            BaseSelector: '#remix-graph-modal-link',
+            BaseSelector: '#remix-graph-button-small',
             Category: programCategory,
             Action: 'click - show remix graph',
             Label: function (e) {
