+++ templates/Program/code_view.html.twig
@@ -1,27 +1,42 @@
-<div class="card">
-  <div class="card-header d-flex justify-content-between expansion-header collapsed" id="codeView"
-       data-toggle="collapse" data-target="#collapseCodeView" aria-expanded="false"
-       aria-controls="collapseCodeView">
-    Code View
-    {% if isVersionSupportedByCatBlocks(version) %}
-      <span id="code-view-pb">{% include 'Default/progressbar.html.twig' with {'progressbar_id': 'code-view-spinner', 'small': true} %}</span>
-    {% endif %}
-    <i id="code-view-toggler" class="material-icons rotate-left {% if isVersionSupportedByCatBlocks(version) %} d-none {% endif %}">chevron_left</i>
-  </div>
+{% extends 'Default/base.html.twig' %}
+{% block page_title %}{{ "codeview.title"|trans({}, "catroweb") }}{% endblock %}
+{% block back_path %}{{ path('program', { 'id' : id }) }}{% endblock %}
 
-  <div id="collapseCodeView"
-       {#  Catblocks needs the container to be visible to generate the SVGs in the correct size #}
-       class="card-body {% if not isVersionSupportedByCatBlocks(version) %} collapse {% endif %}"
-       aria-labelledby="codeView"
-       data-parent="#programDetails">
-    {% if isVersionSupportedByCatBlocks(version) %}
-      <div id='catblocks-code-container'></div>
-      {# As long as catblocks is not stable we need to also load the old view as backup #}
-      {{ render(controller('App\\Catrobat\\Controller\\Web\\CodeViewController:oldView', {'id' : id, 'visible' : false })) }}
-    {% else %}
-      {{ render(controller('App\\Catrobat\\Controller\\Web\\CodeViewController:oldView', {'id' : id })) }}
-    {% endif %}
+{% block head %}
+  <link rel="stylesheet" href="{{ asset('css/'~theme()~'/card.css') }}"/>
 
-  </div>
+  {% if isVersionSupportedByCatBlocks(version) %}
+    <link rel="stylesheet" href="{{ asset('catblocks/main.css') }}"/>
+  {% endif %}
+{% endblock %}
 
-</div>
+{% block body %}
+
+  {% if isVersionSupportedByCatBlocks(version) %}
+    <div id='catblocks-code-container'></div>
+    {# As long as catblocks is not stable we need to also load the old view as backup #}
+    {{ render(controller('App\\Catrobat\\Controller\\Web\\CodeViewController:oldView', {'id' : id, 'visible' : false })) }}
+  {% else %}
+    {{ render(controller('App\\Catrobat\\Controller\\Web\\CodeViewController:oldView', {'id' : id })) }}
+  {% endif %}
+
+{% endblock %}
+
+
+{% block js %}
+
+<script src="{{ asset('js/OldCodeView.min.js') }}"></script>
+
+{% if isVersionSupportedByCatBlocks(version) %}
+  <script src="{{ asset('catblocks/CatBlocks.js') }}"></script>
+  <script src="{{ asset('js/ProgramCodeView.min.js') }}"></script>
+  <script>
+    let codeView = new ProgramCodeView('{{ app.request.getSchemeAndHttpHost() ~ '/' }}',
+      '{{ extracted_path }}', '{{ extracted_dir_hash }}',
+      '{{ app.request.getLocale() }}', '{{ app_env }}', CatBlocks)
+    codeView.init()
+  </script>
+{% endif %}
+
+{% endblock %}
+
