+++ templates/Program/program_apk_generation_buttons.html.twig
@@ -1,0 +1,28 @@
+<div>
+  <button id="apk-generate{{ suffix|default('') }}"
+          class="btn btn-outline-primary btn-block">
+    <i class="material-icons align-bottom">casino</i>
+    <span>{{ "apk.prepare"|trans({}, "catroweb") }}</span>
+  </button>
+</div>
+
+<div>
+  <button id="apk-pending{{ suffix|default('') }}"
+          class="btn btn-outline-primary btn-block text-left d-none">
+    <i class="material-icons align-bottom">
+      {% include 'Default/progressbar.html.twig' with {'progressbar_id': 'apk-progressbar' ~ suffix|default(), 'small': true} %}
+    </i>
+    <span>{{ "apk.preparing"|trans({}, "catroweb") }}</span>
+  </button>
+</div>
+
+<div>
+  <button id="apk-download{{ suffix|default('') }}"
+          class="btn btn-outline-primary btn-block text-left d-none">
+    <i class="material-icons d-none">
+      {% include 'Default/progressbar.html.twig' with {'progressbar_id': 'download-spinner' ~ suffix|default(), 'small': true} %}
+    </i>
+    <i class="material-icons align-bottom">casino</i>
+    <span>{{ "apk.download"|trans({}, "catroweb") }}</span>
+  </button>
+</div>
