+++ templates/Program/program_details.html.twig
@@ -1,0 +1,44 @@
+<div id="details">
+
+  <h2 class="mt-3">
+    {{ "programDetails"|trans({}, "catroweb") }}
+  </h2>
+
+  <div class="row">
+    <div class="col-6 col-lg-4 mt-2 mb-2 no-overflow">
+      <i class="material-icons">visibility</i>
+      <span class="icon-text ml-2 align-bottom">{{ "programs.views"|trans({ '%views%' : program_details.views }, "catroweb") }}</span>
+    </div>
+    <div class="col-6 col-lg-4 mt-2 mb-2 no-overflow">
+      <i class="material-icons">query_builder</i>
+      <span class="icon-text ml-2 align-bottom">{{ program_details.age }}</span>
+    </div>
+    <div class="d-none d-lg-block col-lg-4">
+      {% include 'Program/program_code_statistics_button.html.twig' %}
+    </div>
+  </div>
+
+  <div class="row">
+    <div class="col-6 col-lg-4 mt-2 mb-2 no-overflow mt-lg-3">
+      <i class="material-icons">description</i>
+      <span class="icon-text ml-2">{{ program_details.filesize }} MB</span>
+    </div>
+    <div class="col-6 col-lg-4 mt-2 mb-2 no-overflow mt-lg-3">
+      <i class="material-icons ">get_app</i>
+      <span class="icon-text ml-2">{{ "programs.downloads"|trans({ '%downloads%' : program_details.downloads }, "catroweb") }}</span>
+    </div>
+    <div class="d-none d-lg-block col-lg-4 mt-2">
+      {% include 'Program/program_code_view_button.html.twig' %}
+    </div>
+  </div>
+
+  <div class="row mb-4">
+    <div class="col-6 mb-3 d-lg-none mt-3">
+      {% include 'Program/program_code_statistics_button.html.twig' with {'suffix': '-small'} %}
+    </div>
+    <div class="col-6 mb-3 d-lg-none mt-3">
+      {% include 'Program/program_code_view_button.html.twig' with {'suffix': '-small'} %}
+    </div>
+  </div>
+
+</div>
