--- templates/Program/program_details_accordion.html.twig
@@ -1,12 +1,0 @@
-<div class="text-dark">
-  <div class="accordion" id="programDetails">
-  <div class="row">
-    <div class="col-12">
-        {{ render(controller('App\\Catrobat\\Controller\\Web\\CodeStatisticsController:view', {'id' : program.id})) }}
-    </div>
-    <div class="col-12">
-        {{ render(controller('App\\Catrobat\\Controller\\Web\\CodeViewController:view', {'id' : program.id, 'version' : program_details.languageVersion})) }}
-    </div>
-  </div>
-  </div>
-</div>
