+++ templates/Program/program_reactions.html.twig
@@ -1,0 +1,94 @@
+{# If you change #project-like-buttons, you need to change it in Program.js in sendProjectLike too #}
+{% if program_details.totalLikeCount is same as(0) %}
+  <div id="project-like-buttons{{ suffix|default('') }}"
+     {% if suffix is not defined %} style="display: inline"{% endif %}>
+    <div class="btn btn-primary btn-round d-inline-flex justify-content-center">
+      <i class="material-icons thumbs-up md-24">
+        thumb_up
+      </i>
+    </div>
+  </div>
+  <span id="project-like-counter{{ suffix|default('') }}" class="like-counter d-none">0</span>
+{% else %}
+  <div id="project-like-buttons{{ suffix|default('') }}"
+     {% if suffix is not defined %} style="display: inline"{% endif %}>
+    {% if constant('App\\Entity\\ProgramLike::TYPE_THUMBS_UP') in program_details.activeLikeTypes %}
+      <div class="btn btn-primary btn-round d-inline-flex justify-content-center align-items-center">
+        <i class="material-icons md-24 thumbs-up">
+          thumb_up
+        </i>
+      </div>
+    {% endif %}
+    {% if constant('App\\Entity\\ProgramLike::TYPE_SMILE') in program_details.activeLikeTypes %}
+      <div class="btn btn-primary btn-round d-inline-flex justify-content-center align-items-center">
+        <i class="material-icons md-24 smile">
+          sentiment_very_satisfied
+        </i>
+      </div>
+    {% endif %}
+    {% if constant('App\\Entity\\ProgramLike::TYPE_LOVE') in program_details.activeLikeTypes %}
+      <div class="btn btn-primary btn-round d-inline-flex justify-content-center align-items-center">
+        <i class="material-icons md-24 love">
+          favorite
+        </i>
+      </div>
+    {% endif %}
+    {% if constant('App\\Entity\\ProgramLike::TYPE_WOW') in program_details.activeLikeTypes %}
+      <div class="btn btn-primary btn-round d-inline-flex justify-content-center align-items-center" id="wow-reaction">
+        <img alt="Wow reaction" id="wow-reaction-img-small" class="wow" src="{{ asset('images/default/wow_white.svg') }}">
+      </div>
+    {% endif %}
+  </div>
+  {% if suffix is defined %}
+    <div class="text-nowrap mt-2 align-bottom">
+  {% endif %}
+  <span id="project-like-counter{{ suffix|default('') }}"
+        class="like-counter">
+      {{ program_details.totalLikeCount | humanFriendlyNumber }}
+    {{ 'programs.reactionsText'|trans({}, 'catroweb') }}
+    </span>
+  {% if suffix is defined %}
+    </div>
+  {% endif %}
+{% endif %}
+
+<div id="project-like-detail{{ suffix|default('') }}">
+  {# Thumbs-Up Button #}
+  <div
+     class="btn btn-primary btn-round d-inline-flex justify-content-center align-items-center
+              {% if constant('App\\Entity\\ProgramLike::TYPE_THUMBS_UP') in program_details.activeUserLikeTypes %} active{% endif %}"
+     data-like-type="{{ constant('App\\Entity\\ProgramLike::TYPE_THUMBS_UP') }}"
+     data-toggle="tooltip" title="{{ 'programs.thumbsUp'|trans({}, 'catroweb') }}">
+    <i class="material-icons md-24">
+      thumb_up
+    </i>
+  </div>
+  {# Smile Button #}
+  <div
+     class="btn btn-primary btn-round d-inline-flex justify-content-center align-items-center
+              {% if constant('App\\Entity\\ProgramLike::TYPE_SMILE') in program_details.activeUserLikeTypes %} active{% endif %}"
+     data-like-type="{{ constant('App\\Entity\\ProgramLike::TYPE_SMILE') }}"
+     data-toggle="tooltip" title="{{ 'programs.smile'|trans({}, 'catroweb') }}">
+            <span class="material-icons md-24">
+              sentiment_very_satisfied
+            </span>
+  </div>
+  {# Love Button #}
+  <div
+     class="btn btn-primary btn-round d-inline-flex justify-content-center align-items-center
+              {% if constant('App\\Entity\\ProgramLike::TYPE_LOVE') in program_details.activeUserLikeTypes %} active{% endif %}"
+     data-like-type="{{ constant('App\\Entity\\ProgramLike::TYPE_LOVE') }}"
+     data-toggle="tooltip" title="{{ 'programs.love'|trans({}, 'catroweb') }}">
+            <span class="material-icons md-24">
+              favorite
+            </span>
+  </div>
+  {# Wow Button #}
+  <div
+     class="btn btn-primary btn-round d-inline-flex justify-content-center align-items-center
+              {% if constant('App\\Entity\\ProgramLike::TYPE_WOW') in program_details.activeUserLikeTypes %} active{% endif %}"
+     data-like-type="{{ constant('App\\Entity\\ProgramLike::TYPE_WOW') }}"
+     data-toggle="tooltip" title="{{ 'programs.wow'|trans({}, 'catroweb') }}">
+    <img alt="Wow reaction" id="wow-reaction-img-small" src="{{ asset('images/default/wow_white.svg') }}">
+  </div>
+</div>
