+++ templates/Program/program_reactions_modal.html.twig
@@ -1,0 +1,51 @@
+<div class="modal fade" id="project-like-modal" tabindex="-1" role="dialog"
+     aria-labelledby="project-like-modal-label" aria-hidden="true">
+  <div class="modal-dialog" role="document">
+    <div class="modal-content">
+      <div class="modal-header">
+        <h5 class="modal-title"
+            id="project-like-modal-la</div>bel">{{ "programs.likes.likersTitle"|trans({}, 'catroweb') }}</h5>
+        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+          <span aria-hidden="true">&times;</span>
+        </button>
+      </div>
+      <div class="modal-body">
+        <ul class="nav nav-tabs" role="tablist">
+          <li class="nav-item">
+            <a class="nav-link active" id="all-tab" data-toggle="tab" href="#all-tab-content" role="tab"
+               aria-controls="all-tab-content"
+               aria-selected="true">{{ 'programs.likes.all' | trans({}, 'catroweb') }}<span></span></a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" id="thumbs-up-tab" data-toggle="tab" href="#thumbs-up-tab-content" role="tab"
+               aria-controls="thumbs-up-tab-content" aria-selected="false">
+              <i class="material-icons md-18 thumbs-up">thumb_up</i><span></span></a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" id="smile-tab" data-toggle="tab" href="#smile-tab-content" role="tab"
+               aria-controls="smile-tab-content" aria-selected="false">
+              <i class="material-icons md-18 smile">sentiment_very_satisfied</i><span></span></a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" id="love-tab" data-toggle="tab" href="#love-tab-content" role="tab"
+               aria-controls="love-tab-content" aria-selected="false"><i class="material-icons md-18 love">favorite</i><span></span></a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" id="wow-tab" data-toggle="tab" href="#wow-tab-content" role="tab"
+               aria-controls="wow-tab-content" aria-selected="false"><img alt="Wow reaction" id="wow-reaction-modal" class="wow" src="{{ asset('images/default/wow_black.svg') }}">
+              <span></span></a>
+          </li>
+        </ul>
+        <div class="tab-content" id="myTabContent">
+          <div class="tab-pane fade show active" id="all-tab-content" role="tabpanel"
+               aria-labelledby="all-tab"></div>
+          <div class="tab-pane fade" id="thumbs-up-tab-content" role="tabpanel"
+               aria-labelledby="thumbs-up-tab"></div>
+          <div class="tab-pane fade" id="smile-tab-content" role="tabpanel" aria-labelledby="smile-tab"></div>
+          <div class="tab-pane fade" id="love-tab-content" role="tabpanel" aria-labelledby="love-tab"></div>
+          <div class="tab-pane fade" id="wow-tab-content" role="tabpanel" aria-labelledby="wow-tab"></div>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
