+++ templates/Program/program_remix_graph_button.html.twig
@@ -1,0 +1,9 @@
+<button id="remix-graph-button{{ suffix|default('') }}"
+        class="btn btn-outline-primary btn-block"
+        onclick="location.href='{{ path('remix_graph', {'id': program.id }) }}'">
+  <i class="material-icons d-none">
+    {% include 'Default/progressbar.html.twig' with {'progressbar_id': 'remix-graph-spinner' ~ suffix|default(), 'small': true} %}
+  </i>
+  <i class="material-icons align-bottom">call_split</i>
+  <span>{{ "remixGraph.showRemixGraph"|trans({}, "catroweb") }}</span>
+</button>
