+++ tests/behat/features/web/project-details/project_download.feature
@@ -12,5 +12,12 @@ Feature: As a visitor I want to be able to download projects
   Scenario: I want to download a project via the button
     When I am on "/app/project/1"
     And I wait for the page to be loaded
-    Then the element ".url-download" should be visible
-    And the element ".url-download" should have a attribute "onclick" with value "program.download("
+    Then the element "#url-download-small" should be visible
+    And the element "#url-download-small" should have a attribute "onclick" with value "program.download("
+
+  Scenario: Clicking the download button should deactivate the download button until download is finished
+    Given I am on "/app/project/1"
+    And I wait for the page to be loaded
+    When I click "#url-download-small"
+    Then the element "#download-progressbar-small" should be visible
+    Then the button "#url-download-small" should be disabled until download is finished
